Make sure all colors work with light/dark mode & save the zoom state of the app
