+++ moodterm/moodtermApp.swift
@@ -7,17 +7,8 @@ struct MoodtermApp: App {
 
     var body: some Scene {
         WindowGroup {
-            TabView(
-                tabs: $appState.tabs, selectedTab: $appState.selectedTab,
-                fontSizeFactor: $appState.fontSizeFactor
-            )
-            .onAppear {
-                appState.selectedTab = appState.tabs.first?.id
-                appState.observeTabChanges()
-            }
-            .onChange(of: appState.tabs) { _ in
-                appState.observeTabChanges()
-            }
+            ContentView()
+                .environmentObject(appState)
         }
         .commands {
             FontSizeCommands(fontSizeFactor: $appState.fontSizeFactor)
@@ -40,33 +31,20 @@ struct MoodtermApp: App {
     }
 }
 
-class AppState: ObservableObject {
-    @Published var tabs: [Tab] = MoodtermApp.loadTabs()
-    @Published var selectedTab: UUID?
-    @Published var fontSizeFactor: Double = 1.0
-    var cancellables = Set<AnyCancellable>()
+struct ContentView: View {
+    @EnvironmentObject var appState: AppState
 
-    func observeTabChanges() {
-        cancellables.removeAll()
-        for tab in tabs {
-            tab.$currentDirectory
-                // Debounce to make sure that the tabs are saved after the currentDirectory has been updated
-                .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
-                .sink { [weak self] _ in
-                    self?.saveTabs(self?.tabs ?? [])
-                }
-                .store(in: &cancellables)
+    var body: some View {
+        TabView(
+            tabs: $appState.tabs, selectedTab: $appState.selectedTab,
+            fontSizeFactor: $appState.fontSizeFactor
+        )
+        .onAppear {
+            appState.selectedTab = appState.tabs.first?.id
+            appState.observeTabChanges()
         }
-    }
-
-    func saveTabs(_ tabs: [Tab]) {
-        do {
-            // print the current directory from the first tab
-            print("Saving directory: \(tabs.first?.currentDirectory ?? "")")
-            let data = try JSONEncoder().encode(tabs)
-            UserDefaults.standard.set(data, forKey: "savedTabs")
-        } catch {
-            print("Failed to save tabs: \(error)")
+        .onChange(of: appState.tabs) {
+            appState.observeTabChanges()
         }
     }
 }
